refactor(plan-detail): extract fetchPlanSnapshot/useInitialFetch/usePolling
Splits the snapshot fetch + lifecycle out of usePlanDetailPage:
- shell/hooks/fetchPlanSnapshot.ts now owns the Connect RPC reads
  (project, currentUser, plan, issue, planCheckRuns, rollout, taskRuns)
  and the convertRouteQuery helper.
- shell/hooks/useInitialFetch.ts owns the mount-time reset + initial
  load effect, including the ConnectError 404/403 routing.
- shell/hooks/usePolling.ts owns the exponential-backoff polling
  timer and its cleanup.

usePlanDetailPage.ts goes from 653 to 412 lines while keeping
behavior, ordering, jitter, and cleanup semantics identical.

Note: the existing fetchPlanDetailSnapshot already used Connect
clients directly; no Pinia store reads were present in this file,
so no app-store slice migration was needed for this task.
